PyRoPy: A Python Roguelike v0.01a
This is a python roguelike that I'll be maintaining as a hobby project.
Parts of this will follow the libtcod tutorial.

UPDATE log
- 0.01a: Bare bones, an @ moves around the screen. Nothing much to see
yet.

DISCLAIMER:
This program is free software: you can redistribute it and/or modify it
under the terms of the GNU General Public License as published by the
Free Software Foundation, either version 3 of the License, or (at your
option) any later version.

This program is distributed in the hope that it will be useful as
educational tool, but WITHOUT ANY WARRANTY; without even the implied
warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See
the GNU General Public License for more details.

You should have received a copy of the GNU General Public License along
with this program.  If not, see <http://www.gnu.org/licenses/>